+++ app/Http/Controllers/Admin/WeeklyDealsController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Models\WeeklyDeal;
 use Illuminate\Http\Request;
 use File;
+use Illuminate\Support\Facades\Gate;
 
 class WeeklyDealsController extends Controller
 {
@@ -16,6 +17,8 @@ class WeeklyDealsController extends Controller
      */
     public function index()
     {
+        abort_unless(Gate::allows('View Deal Of The Week'), 403);
+
         $deals = WeeklyDeal::latest()->paginate(10);
         createLog('viewed weekly deals details'); // activity log
 
@@ -29,6 +32,8 @@ class WeeklyDealsController extends Controller
      */
     public function create()
     {
+        abort_unless(Gate::allows('Create Deal Of The Week'), 403);
+
         return view('admin.weekly-deals.create');
     }
 
@@ -40,6 +45,8 @@ class WeeklyDealsController extends Controller
      */
     public function store(Request $request)
     {
+        abort_unless(Gate::allows('Create Deal Of The Week'), 403);
+
         $input = $request->except('image');
         $input['image'] = $this->fileUpload($request, 'image');
         WeeklyDeal::create($input);
@@ -67,6 +74,8 @@ class WeeklyDealsController extends Controller
      */
     public function edit(WeeklyDeal $deal)
     {
+        abort_unless(Gate::allows('Edit Deal Of The Week'), 403);
+
         return view('admin.weekly-deals.edit', compact('deal'));
     }
 
@@ -79,6 +88,8 @@ class WeeklyDealsController extends Controller
      */
     public function update(Request $request, WeeklyDeal $deal)
     {
+        abort_unless(Gate::allows('Edit Deal Of The Week'), 403);
+
         $input = $request->except('image');
         $image = $this->fileUpload($request, 'image');
         if ($image) {
@@ -100,6 +111,8 @@ class WeeklyDealsController extends Controller
      */
     public function destroy(WeeklyDeal $deal)
     {
+        abort_unless(Gate::allows('Delete Deal Of The Week'), 403);
+
         $this->removeFile($deal->image);
         $deal->delete();
         createLog('deleted a weekly deals'); // activity log
